Added documentation for Testbench
changed files included in build so that they can be run without need for modifications. (browser environments same as defined in grid)

Removed enter character from TestConverter so that it is run as a selenium doCommand, this makes the execution much faster and lessens the communication between hub and rc.

Added appendToTest command for connecting tests together on conversion.

Fixed vaadin-format.js

git-svn-id: http://dev.vaadin.com/svn/TestingTools/2.0@9651 de684a11-831f-0410-ab27-cb86445eb682