property exclude.runner will exclude TestBenchRunner  and its Api documentation from the build.
Changed IDE name Test Case to Test. This due to TestBench now having only Tests and Test Suites.

Updated Practices documentation and renamed it.
Updated TestBench documentation to use Test and not Test Case.

Set IDE to record xpath for uriFragment links so that mouseClick works. This could/should be changed when a mouseClick with the Vaadin locator works.

git-svn-id: http://dev.vaadin.com/svn/TestingTools/2.0@10734 de684a11-831f-0410-ab27-cb86445eb682