Use Javassist for element and driver proxies (#13159)
Add proxies to work with Elements provided by TestBench.
This is needed to fix broken Actions in TestBench4. 

Provide a fast new MethodHandler to call waitForVaadin while 
doing things with the driver. 

Implement MethodFilter and MethodHandler for Elements so calls
to Selenium WebElement functionality not directly implemented in 
TestBench gets executed the right way.

Change-Id: I2fe199aec1413dd005d576cd2622536abc5e1964