fix: Made withTextContaining work with all components
Currently withTextContaining returns false if the component is not HasText,
Html or if element is not a text node.
This change adds support for all other component types, by checking the input
against Element.getTextRecursively.
To be noted that the result from getTextRecursively() may be a the concatenation
of text from child nodes, but without space separators