feat: log all exceptions caught by RetryRule
Currently, RetryRule show details only of the last caught exception, but this
one may not be the real reason of the test failure.
This change prints the stack trace of all caught expections when debug level
for RetryRule is at least 'debug', helping the develper to detect the real
cause of the test failure.

Fixes #1527