fix: override remote driver default timeouts
By default remote driver is created with a connection timeout of 2 minutes
and a read timeout of 3 hours. This sometimes causes the validation jobs
to fail with execution timeout, because the remote driver client hangs
because of the 3 hours read timeout.
This change resets the timeout values to be similar to the configuration
used by recent selenium versions.