fix: make tabs tester index operations work on visible tabs
The Tabs tester index base operation currently does not care about
tabs visibility, thus not correctly reflecting the action in the browser.
If there are three tabs, but the one in the middle is hidden, operation
for index 2 should effectively act on the third registered tab.
This change fixes fetching the tab by index by filtering out hidden tabs.