Switch to use Vaadin 8.29+
This version of TestBench does not support the free Vaadin 8 versions
out of the box, because some API changes from upgrading to Selenium 4
are not backwards compatible. These changes will be reflected in
vaadin-testbench-api 8.29.0 and newer, requiring a commercial
subscription.

Because Vaadin 8.29.0 hasn't been released yet, and can't be before
there is a TestBench release matching the new API, the initial
dependency is to the already available snapshot release. The dependency
should be updated to a stable release once it becomes available.